Move database into lib
Follow up to #108
