+++ test/integration/post-resource/index.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const request = require('supertest');
 const app = require('../../../server/app');
-const getDb = require('../../../database');
+const getDb = require('../../../lib/database');
 const wipeDatabase = require('../../../lib/wipe-database');
 const validators = require('../../helpers/json-api-validators');
