+++ test/integration/setup.js
@@ -1,4 +1,4 @@
-const configureDb = require('../../database');
+const configureDb = require('../../lib/database');
 
 // We need to set up the database to be in the test environment.
 configureDb({test: true});
